requirements.txt
vk_api>=11.8.0,<12.0.0
python-dotenv>=0.13.0,<0.14.0
pyowm>=2.10.0,<2.11.0